Move TestStatusRepository to test fixtures
COAND-911
