+++ components-core/src/testFixtures/java/com/adyen/checkout/components/core/internal/data/api/TestStatusRepository.kt
@@ -1,29 +1,21 @@
 /*
- * Copyright (c) 2022 Adyen N.V.
+ * Copyright (c) 2024 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by oscars on 15/8/2022.
+ * Created by josephj on 19/7/2024.
  */
 
-package com.adyen.checkout.components.core.internal.test
+package com.adyen.checkout.components.core.internal.data.api
 
-import androidx.annotation.RestrictTo
-import com.adyen.checkout.components.core.internal.data.api.StatusRepository
 import com.adyen.checkout.components.core.internal.data.model.StatusResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.asFlow
 
-/**
- * Test implementation of [StatusRepository]. This class should never be used in not test code as it does not actuall
- * poll any status!
- */
-// TODO move to test fixtures once it becomes supported on Android
-@RestrictTo(RestrictTo.Scope.TESTS)
 class TestStatusRepository : StatusRepository {
 
     var pollingResults: List<Result<StatusResponse>> = emptyList()
-    var timesOnPollCalled: Int = 0
+    private var timesOnPollCalled: Int = 0
 
     override fun poll(paymentData: String, maxPollingDuration: Long): Flow<Result<StatusResponse>> {
         timesOnPollCalled++
